coverity: fix upgrade releated issues [1/1]
PD#SWPL-97255

Problem:
coverity repport issues

Solution:
fix tained string issue in cmd_imgread/cmd_bcb/efuse)usr_space_api
fix goto pass through issue in cmd_imgred
fix overrun in aml_dnl/v3_tool_simg2img
fix tained string on f_v3_usb_tool

Verify:
t5w_at301

+++ drivers/usb/gadget/v3_burning/v3_common/v3_tool_simg2img.c
@@ -420,7 +420,7 @@ int v3tool_simg2img_verify_img(sha1_context* ctx, const char* partName, int64_t 
     for ( ; _iChunk < nChunk; ++_iChunk, ++backInf )
     {
         if (!_dataChunkLeft) sha1_update(ctx, (u8*)backInf, CHUNK_HEAD_SIZE);
-        const int64_t flashSpace = backInf->chunk_sz * _spPacketStates.sparseBlkSz;
+        const int64_t flashSpace = backInf->chunk_sz * (int64_t)_spPacketStates.sparseBlkSz;
         switch (backInf->chunk_type)
         {
             case CHUNK_TYPE_RAW:
